LayoutTests/inspector/profiler/heap-snapshot-reveal-in-dominators-view-expected.txt

Tests that an object can be revealed in the Dominators view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll
Revealing node...
Node revealed.

Profiler was disabled.
